Add DCMI "Get Power Reading" and "Get Asset Tag"
This commit adds support for two commands from the Intel DCMI [1]
extenstions for IPMI. DCMI uses the Net Function "Group Extension", ID
`0xDC`. Two commands implemented in the library, `GetDCMIPowerReading`
and `GetDCMIAssetTag`. The corresponding commands are added to the CLI
tool and the documentation in the README.

Tested against SuperMicro hardware.

[1] https://www.intel.com/content/dam/www/public/us/en/documents/technical-specifications/dcmi-v1-5-rev-spec.pdf